api/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pg8000==1.30.3
python-dotenv==1.0.1
gunicorn==21.2.0 